+++ services/usage-service/src/main/java/com/eevee/usageservice/api/dto/UsageLogApiKeyItemResponse.java
@@ -1,0 +1,4 @@
+package com.eevee.usageservice.api.dto;
+
+public record UsageLogApiKeyItemResponse(String apiKeyId) {
+}
