+++ services/usage-service/src/main/java/com/eevee/usageservice/api/dto/UsageLogEntryResponse.java
@@ -9,6 +9,7 @@ public record UsageLogEntryResponse(
         Instant occurredAt,
         String correlationId,
         String provider,
+        String apiKeyId,
         String model,
         Long promptTokens,
         Long completionTokens,
